requirements.txt
annotated-doc
annotated-types
anyio
attrs
backoff
bcrypt
build
cachetools
certifi
charset-normalizer
chromadb
click
coloredlogs
contourpy
cycler
distro
durationpy
fastapi
filelock
filetype
flatbuffers
fonttools
fsspec
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
grpcio
grpcio-status
h11
hf-xet
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
kubernetes
langchain
langchain-chroma
langchain-core
langchain-google-genai
langchain-huggingface
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
markdown-it-py
MarkupSafe
matplotlib
mdurl
mmh3
mpmath
networkx
numpy
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
pandas
pillow
posthog
proto-plus
protobuf
pyasn1
pyasn1_modules
pybase64
pydantic
pydantic_core
Pygments
pyparsing
PyPika
pyproject_hooks
python-dateutil
python-dotenv
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
starlette
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
transformers
typer
typing-inspection
typing_extensions
tzdata
urllib3
uuid_utils
uvicorn
uvloop
watchfiles
websocket-client
websockets
xxhash
zipp
zstandard
